[FMGEN] Maybe FIX FTBFS with some compiler (at least not with CLANG++-13). See https://matsuri.5ch.net/test/read.cgi/i4004/1601147054/451.

+++ source/src/vm/fmgen/misc.h
@@ -25,7 +25,7 @@ inline int Limit(int v, int max, int min)
 #if defined(__has_builtin) && (__has_builtin(__builtin_bswap32))
 inline unsigned int BSwap(unsigned int a)
 {
-	return __builtin_bswap32(a);
+	return (unsigned int)__builtin_bswap32((uint32_t)a);
 }
 #else
 inline unsigned int BSwap(unsigned int a)
